requirements_test.txt
django==4.2.19
psycopg2-binary
gunicorn==23.0.0
django-cors-headers==4.7.0
requests==2.32.3
djangorestframework==3.15.2
django-auth-adfs==1.15.0
setuptools==75.8.0
azure-storage-blob==12.24.1
whitenoise==6.9.0
urllib3>=2.2.0,<3.0.0
Pillow==11.1.0
python-magic==0.4.27
django-ratelimit==4.1.0
packaging>=24.0
